// components/ui/sidebar/types.ts
import type { SheetContent } from "@/components/ui/sheet";
export type SidebarContextProps = {
  state: "expanded" | "collapsed";
  open: boolean;
  setOpen: (open: boolean) => void;
  openMobile: boolean;
  setOpenMobile: (open: boolean) => void;
  isMobile: boolean;
  toggleSidebar: () => void;
};
export type SidebarProps = React.ComponentProps<"div"> & {
  side?: "left" | "right";
  variant?: "sidebar" | "floating" | "inset";
  collapsible?: "offcanvas" | "icon" | "none";
  mobileTitle?: string;
  mobileDescription?: string;
  mobileContentProps?: React.ComponentProps<typeof SheetContent>;
};